@@ -1,4 +1,5 @@
 [metadata]
+# https://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
 name = edapy
 
 author = Martin Thoma
@@ -31,6 +32,8 @@ classifiers =
     Natural Language :: English
     Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
     Topic :: Scientific/Engineering :: Information Analysis
     Topic :: Software Development
     Topic :: Utilities
@@ -38,11 +41,11 @@ classifiers =
 [options]
 zip_safe = false
 packages = find:
-python_requires = >= 3.0
+python_requires = >=3.6
 
 [options.entry_points]
 console_scripts =
-    hwrt = edapy.cli:entry_point
+    edapy = edapy.cli:entry_point
 
 [tool:pytest]
 addopts = --doctest-modules --mccabe --cov=./edapy --cov-append --cov-report html:tests/reports/coverage-html --cov-report xml:tests/reports/coverage.xml --cov-report term --flake8 --ignore=docs/ --durations=3 --black
@@ -50,17 +53,22 @@ doctest_encoding = utf-8
 mccabe-complexity=12
 
 [pydocstyle]
-convention=numpy
 match_dir = mpu
+ignore = D105, D413, D107, D416, D212, D203, D417
 
 [flake8]
-exclude = tests/
-ignore = H301, H306, H404, H405, W503, E501, E203, D413
+max_line_length = 88
+exclude = tests/*,.tox/*,.nox/*
+ignore = H301,H306,H404,H405,W503,E203,D413
 
 [mutmut]
 backup=False
 runner=python3 -m pytest
 tests_dir=tests/
+
+
+[mypy]
+ignore_missing_imports = True
 
 [coverage:run]
 parallel=true
