@@ -2,9 +2,10 @@
 
 # Core Library
 import logging
+from typing import Any, Dict
 
 
-def pytest_configure(config):
+def pytest_configure(config: Dict[Any, Any]) -> None:
     """Flake8 is to verbose. Mute it."""
     logging.getLogger("flake8").setLevel(logging.WARN)
     logging.getLogger("pydocstyle").setLevel(logging.INFO)
